+++ src/main/java/com/umc/puppymode2/domain/user/auth/dto/LoginResponseDTO.java
@@ -1,4 +1,4 @@
-package com.umc.puppymode2.domain.user.dto;
+package com.umc.puppymode2.domain.user.auth.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
