+++ src/main/java/com/umc/puppymode2/domain/user/auth/enums/Provider.java
@@ -1,0 +1,5 @@
+package com.umc.puppymode2.domain.user.auth.enums;
+
+public enum Provider {
+    KAKAO
+}
