--- src/main/java/com/umc/puppymode2/global/auth/enums/Provider.java
@@ -1,5 +1,0 @@
-package com.umc.puppymode2.global.auth.enums;
-
-public enum Provider {
-    KAKAO
-}
